apps/processor/requirements.txt
kafka-python
qdrant-client
face_recognition
Pillow
numpy
requests
fastapi
uvicorn
python-multipart
